Kobo writer: additional keywords as alternatives
Before this commit, multiple keywords for a single entry were just
joined with a pipe symbol. This is not what you want in a kobo
dictionary. This commit uses the first keyword as the headword for the
entry and adds the remaining ones as alternatives.

Open TODO:
* If the alternatives don't have the same prefix, they have to be
copied into each dicthtml file. This should probably be done as a
preprocessing step before sorting the dictionary.