add mobi7-forcing switch to kindlegen command
This will speed up conversion to mobi by skipping kindlegen's attempts to create an enhanced mobi (KF8), because Kindle dictionaries cannot be made in the KF8 format anyway. Kindlegen does automatically detect that it must create a mobi7 file, but not until after it's parsed some number of the source xhtml files listed in content.opf.